Rename login to cadastro
Organize function structure
